changed DP_Improved
pushing changes that include what was discussed on zoom.

Changed naming conventions for data structures.

Currently identified a problem with the subtree seed allocation where we are assuming that we are NOT seeding the root, so all k seeds are allocated among the subtrees.

However, we must also consider seeding the root, which means we must also consider all the partitions of k-1 seeds along with the computation of negative payoff from the root-child edges.

All in all, I am going to try to write code to reflect this problem and take a deeper look at the DP algorithm.